requirements.txt
Flask
flask-socketio
eventlet                # or gevent
kafka-python            # or confluent-kafka
openai
python-dotenv
